# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ShipSmart API",
        "description": "API backend para teste técnico de TechLead  na ShipSmart",
        "contact": {
            "name": "Vitor Pedroso",
            "email": "vitorsamuel283@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "List all contacts (paginated)",
                "operationId": "bc39494d59500e2d75e6f6648b43c8de",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (default: 10)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of contacts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "nome": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "telefone": {
                                                        "type": "string"
                                                    },
                                                    "cep": {
                                                        "type": "string"
                                                    },
                                                    "estado": {
                                                        "type": "string"
                                                    },
                                                    "cidade": {
                                                        "type": "string"
                                                    },
                                                    "bairro": {
                                                        "type": "string"
                                                    },
                                                    "endereco": {
                                                        "type": "string"
                                                    },
                                                    "numero": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Create a new contact",
                "operationId": "779a1f2268bae1138e7ea7843dd88d90",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "cep"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Vitor Pedroso"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "vitor@example.com"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "11999999999"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "01001000"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "SP"
                                    },
                                    "cidade": {
                                        "type": "string",
                                        "example": "São Paulo"
                                    },
                                    "bairro": {
                                        "type": "string",
                                        "example": "Sé"
                                    },
                                    "endereco": {
                                        "type": "string",
                                        "example": "Praça da Sé"
                                    },
                                    "numero": {
                                        "type": "string",
                                        "example": "100"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/contacts/{id}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Retrieve a contact by ID",
                "operationId": "2a2c193f37e667891b7a6bc95e8e4132",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact details"
                    },
                    "404": {
                        "description": "Contact not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Update an existing contact",
                "operationId": "45636eb97c78bd0e474d5507fde18f1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "cep"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Vitor Pedroso Updated"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "vitor.updated@example.com"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "11988888888"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "01310930"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "SP"
                                    },
                                    "cidade": {
                                        "type": "string",
                                        "example": "São Paulo"
                                    },
                                    "bairro": {
                                        "type": "string",
                                        "example": "Bela Vista"
                                    },
                                    "endereco": {
                                        "type": "string",
                                        "example": "Avenida Paulista"
                                    },
                                    "numero": {
                                        "type": "string",
                                        "example": "1578"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contact updated successfully"
                    },
                    "404": {
                        "description": "Contact not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Delete a contact by ID",
                "operationId": "a4fa6d0e82512d1fac868a7b61e6a2e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Contact deleted successfully"
                    },
                    "404": {
                        "description": "Contact not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Contacts",
            "description": "Endpoints for managing contacts"
        }
    ]
}